Season 2 - Material Soul ratios + algorithm adjustments (#315)
* Adjusted SP ratios for materials
* Adjusted logic

+++ pallets/ajuna-awesome-avatars/src/types/avatar/versions/v2/test_utils.rs
@@ -131,7 +131,7 @@ pub(crate) fn create_random_blueprint(
 					slot_type,
 					equippable_type,
 					material_pattern,
-					soul_points,
+					soul_points as u8,
 				)
 				.build()
 		}),
